Conform serialization to ngen BMI serialization protocol
Fixing conformance with ngen protocol:

- add include/ngen_utilities.h holding the reserved name, type, unit and
  item count tables, indexed by a shared enum
- rename involved variables to have `ngen::` prefix
- resolve the variables in Get_var_units, which is the engine's entire
  support probe and previously could never find them
- tweak types of serialization_size and serialization_create to int
- keep the variables out of output_var_names and input_var_names, and
  leave Get_var_grid and Get_var_location failing for them
- don't fail Get_var_nbytes or Get_value_ptr during restore, when they
  are needed but no buffer exists yet

+++ src/bmi_serialization.cpp
@@ -1,6 +1,7 @@
 extern "C" {
 #include "../include/topmodel.h"
 }
+#include <cstdint>
 #include <cstdio>
 #include "../include/bmi_serialization.h"
 
@@ -129,7 +130,7 @@ const int serialize_topmodel(Bmi* bmi) {
     }
     // set size and allocate memory
     uint64_t serialized_size = stream.size();
-    model->serialized_length = serialized_size + sizeof(uint64_t);
+    model->serialized_length = (int)(serialized_size + sizeof(uint64_t));
     model->serialized = (char*)malloc(model->serialized_length);
     // make sure memory could be allocated
     if (model->serialized == NULL) {
